+++ msalautomationapp/src/androidTest/java/com/microsoft/identity/client/msal/automationapp/testpass/broker/ltw/TestCase2572294.java
@@ -20,11 +20,12 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.client.msal.automationapp.testpass.msalonly.ltw;
+package com.microsoft.identity.client.msal.automationapp.testpass.broker.ltw;
 
 import com.microsoft.identity.client.msal.automationapp.R;
 import com.microsoft.identity.client.msal.automationapp.testpass.broker.AbstractMsalBrokerTest;
 import com.microsoft.identity.client.ui.automation.annotations.LTWTests;
+import com.microsoft.identity.client.ui.automation.annotations.SupportedBrokers;
 import com.microsoft.identity.client.ui.automation.app.MsalTestApp;
 import com.microsoft.identity.client.ui.automation.broker.BrokerCompanyPortal;
 import com.microsoft.identity.client.ui.automation.broker.BrokerLTW;
@@ -37,22 +38,23 @@ import com.microsoft.identity.labapi.utilities.constants.TempUserType;
 import org.junit.Assert;
 import org.junit.Test;
 
-// Authenticator has highest priority  - Case4 (Auth, CP, LTW)
-// https://identitydivision.visualstudio.com/Engineering/_workitems/edit/2584411
+// Authenticator has highest priority  - Case1 (LTW, Auth, CP)
+// https://identitydivision.visualstudio.com/Engineering/_workitems/edit/2572294
 @LTWTests
-public class TestCase2584411 extends AbstractMsalBrokerTest {
+@SupportedBrokers(brokers = {BrokerLTW.class})
+public class TestCase2572294 extends AbstractMsalBrokerTest {
     @Test
-    public void test_2584411() throws Throwable {
+    public void test_2572294() throws Throwable {
         final String username = mLabAccount.getUsername();
         final String password = mLabAccount.getPassword();
+
+        final BrokerMicrosoftAuthenticator brokerMicrosoftAuthenticator = new BrokerMicrosoftAuthenticator();
+        brokerMicrosoftAuthenticator.uninstall();
+        brokerMicrosoftAuthenticator.install();
 
         final BrokerCompanyPortal brokerCompanyPortal = new BrokerCompanyPortal();
         brokerCompanyPortal.uninstall();
         brokerCompanyPortal.install();
-
-        final BrokerLTW brokerLTW = new BrokerLTW();
-        brokerLTW.uninstall();
-        brokerLTW.install();
 
         final MsalTestApp msalTestApp = new MsalTestApp();
         msalTestApp.install();
@@ -86,28 +88,28 @@ public class TestCase2584411 extends AbstractMsalBrokerTest {
         Assert.assertEquals("Active broker pkg name : " + BrokerMicrosoftAuthenticator.AUTHENTICATOR_APP_PACKAGE_NAME, activeBroker);
     }
 
-    @Override
-    public LabQuery getLabQuery() {
-        return null;
-    }
+        @Override
+        public LabQuery getLabQuery() {
+            return null;
+        }
 
-    @Override
-    public TempUserType getTempUserType() {
-        return TempUserType.BASIC;
-    }
+        @Override
+        public TempUserType getTempUserType() {
+            return TempUserType.BASIC;
+        }
 
-    @Override
-    public String[] getScopes() {
-        return new String[]{"User.read"};
-    }
+        @Override
+        public String[] getScopes() {
+            return new String[]{"User.read"};
+        }
 
-    @Override
-    public String getAuthority() {
-        return mApplication.getConfiguration().getDefaultAuthority().getAuthorityURL().toString();
-    }
+        @Override
+        public String getAuthority() {
+            return mApplication.getConfiguration().getDefaultAuthority().getAuthorityURL().toString();
+        }
 
-    @Override
-    public int getConfigFileResourceId() {
-        return R.raw.msal_config_default;
-    }
+        @Override
+        public int getConfigFileResourceId() {
+            return R.raw.msal_config_default;
+        }
 }
